requirements.txt
Pillow
einops
safetensors
timm
tomesd
torch>=2.1.2
torchdiffeq
torchsde
xformers
decord
datasets
numpy
scikit-image
opencv-python-headless
omegaconf
SentencePiece
albumentations
imageio[ffmpeg]
imageio[pyav]
tensorboard
beautifulsoup4
ftfy
func_timeout
deepspeed
accelerate>=0.25.0
gradio>=3.41.2,<=3.48.0
diffusers>=0.30.1
transformers>=4.37.2
nvitop
came-pytorch
open_clip_torch
pytorch_fid
lpips
ffmpeg-python